Transacciones anadidas al ABM receta, conexion con el microservicio receta

+++ Application/UseCases/Receta/RecetaService.cs
@@ -3,13 +3,8 @@ using Application.Interfaces;
 using Application.Mappers;
 using Application.Request;
 using Application.Response;
-using Azure.Core;
 using Domain.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.Transactions;
 
 namespace Application.UseCases.SReceta
 {
@@ -36,10 +31,12 @@ namespace Application.UseCases.SReceta
         //------ Metodos ABM ------
         public async Task<RecetaResponse> CreateReceta(RecetaRequest recetaRequest)
         {
+            using var transactionScope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled);
+
             try
             {
                 Receta recetaCreada = null;
-                if(await VerifyAll(recetaRequest))
+                if (await VerifyAll(recetaRequest))
                 {
                     TimeSpan tiempoPreparacion = await GetHorario(recetaRequest.TiempoPreparacion);
                     Receta receta = new Receta
@@ -53,21 +50,13 @@ namespace Application.UseCases.SReceta
                         TiempoPreparacion = tiempoPreparacion,
                         IngredentesReceta = new List<IngredienteReceta>(),
                         Pasos = new List<Paso>(),
-                };
+                    };
                     recetaCreada = await _command.CreateReceta(receta);
                 }
-                //recetaRequest.ListaPasos.ForEach(paso => _pasoService.CreatePaso(paso, recetaCreada.RecetaId));
-                foreach (PasoRequest paso in recetaRequest.ListaPasos)
-                {
-                    await _pasoService.CreatePaso(paso, recetaCreada.RecetaId);
-                }
-                //recetaRequest.ListaIngredienteReceta.ForEach(ingReceta => _ingRecetaService.CreateIngredienteReceta(ingReceta, recetaCreada.RecetaId));
-                foreach (IngredienteRecetaRequest ingReceta in recetaRequest.ListaIngredienteReceta)
-                {
-                    await _ingRecetaService.CreateIngredienteReceta(ingReceta, recetaCreada.RecetaId);
-                }
-                //return await CreateRecetaResponse(await _query.GetRecetaById(recetaCreada.RecetaId));
-                return await _recetaMapper.CreateRecetaResponse(await _query.GetRecetaById(recetaCreada.RecetaId));
+                foreach (PasoRequest paso in recetaRequest.ListaPasos) { await _pasoService.CreatePaso(paso, recetaCreada.RecetaId); }
+                foreach (IngredienteRecetaRequest ingReceta in recetaRequest.ListaIngredienteReceta) { await _ingRecetaService.CreateIngredienteReceta(ingReceta, recetaCreada.RecetaId); }
+                transactionScope.Complete();
+                return await _recetaMapper.CreateRecetaResponse(recetaCreada);
             }
             catch (ExceptionSintaxError e)
             {
@@ -91,33 +80,25 @@ namespace Application.UseCases.SReceta
             {
                 //Ver como hacer para que se muestren los pasos de dicha receta
                 //En un futuro agregar un validador de urls con eso de las fotos. Pero más adelante!!
+                using var transactionScope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled);
                 Receta unaReceta = null;
-                if (! await VerifyRecetaId(recetaId)) { throw new ExceptionNotFound("El id de la receta no existe"); }
+                if (!await VerifyRecetaId(recetaId)) { throw new ExceptionNotFound("El id de la receta no existe"); }
                 if (await VerifyAll(recetaRequest))
                 {
                     //Acá hay que verificar los conflictos posibles (qué conflictos podría haber...)
                     //Posibles conflictos: sin ideas
                     TimeSpan tiempoPreparacion = await GetHorario(recetaRequest.TiempoPreparacion);
                     unaReceta = await _command.UpdateReceta(recetaRequest, recetaId);
-                    //Agregarlo a un método privado 
-                    //ver a nivel logico si vamos a hacer que se traiga en el request todo, o solo lo que se va a modificar
-                    foreach(IngredienteRecetaRequest ingReceta in recetaRequest.ListaIngredienteReceta)
+                    if (!await _ingRecetaService.DeleteAllIngRecetaWhitRecetaId(recetaId) && ! await _pasoService.DeleteAllPasosWhitRecetaId(recetaId))
                     {
-                        //capaz pongo un bool true o false
-                        //Necesito ver la forma de reemplazar uno con el otro
-                        int ingRecetaId = await _ingRecetaService.GetIngredienteRecetaBy(recetaId,ingReceta.ingredienteId);
-                        IngredienteRecetaResponse pepe = await _ingRecetaService.UpdateIngredienteReceta(ingReceta, ingRecetaId); 
+                        foreach (PasoRequest paso in recetaRequest.ListaPasos) { await _pasoService.CreatePaso(paso, recetaId); }
+                        foreach (IngredienteRecetaRequest ingReceta in recetaRequest.ListaIngredienteReceta) { await _ingRecetaService.CreateIngredienteReceta(ingReceta, recetaId); }
                     }
-                    foreach (PasoRequest pasoRequest in recetaRequest.ListaPasos)
-                    {
-                        //Hacer lo mismo que hice con la otra lsita xD
-                        _pasoService.GetPasoidByRecetaId(recetaId, pasoRequest.Orden);
-                        PasoResponse pepo = await _pasoService.UpdatePaso(pasoRequest, 2);
-                    }
-                    
+
                 }
-                return await CreateRecetaResponse(unaReceta);
-            } 
+                transactionScope.Complete();
+                return await _recetaMapper.CreateRecetaResponse(unaReceta);
+            }
             catch (Conflict ex)
             {
                 throw new Conflict("Error en la implementación a la base de datos: " + ex.Message);
@@ -139,8 +120,14 @@ namespace Application.UseCases.SReceta
                 //Hay que ver a nivel de usuario como se crea todo esto porque va de la mano con lo de usuario :O
                 //Cuando se borre la receta tenemos que implementar que se borren todos los pasos <----- (De acá vendría el conflict)
                 //Tambien que se borren los ingredientereceta 
+                using var transactionScope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled);
+                Receta recetaToDelete = null;
                 if (!await VerifyRecetaId(id)) { throw new ExceptionNotFound("El id no existe"); }
-                Receta recetaToDelete = await _command.DeleteReceta(await _query.GetRecetaById(id));
+                if (!await _ingRecetaService.DeleteAllIngRecetaWhitRecetaId(id) && !await _pasoService.DeleteAllPasosWhitRecetaId(id))
+                {
+                    recetaToDelete = await _command.DeleteReceta(await _query.GetRecetaById(id));
+                }
+                transactionScope.Complete();
                 return new RecetaDeleteResponse
                 {
                     id = recetaToDelete.RecetaId,
@@ -164,7 +151,7 @@ namespace Application.UseCases.SReceta
                 var paso = await _query.GetRecetaById(id);
                 if (paso != null)
                 {
-                    return await CreateRecetaResponse(paso);
+                    return await _recetaMapper.CreateRecetaResponse(paso);
                 }
                 else
                 {
@@ -191,73 +178,15 @@ namespace Application.UseCases.SReceta
 
             foreach (var receta in recetas)
             {
-                recetasResponse.Add(await CreateRecetaResponse(receta));
+                recetasResponse.Add(await _recetaMapper.CreateRecetaResponse(receta));
             }
             return recetasResponse;
         }
 
         //----- Métodos privados -----
 
-        //----- Generadores de responses ------
-        private async Task<RecetaResponse> CreateRecetaResponse(Receta unaReceta)
-        {
-            var receta = new RecetaResponse
-            {
-                Id = unaReceta.RecetaId,
-                //Categoria = unaReceta.CategoriaRecetaId,
-                Dificultad = await CreateDificultadResponse(unaReceta.Dificultad),
-                FotoReceta = unaReceta.FotoReceta,
-                TiempoPreparacion = unaReceta.TiempoPreparacion.ToString(),
-                Titulo = unaReceta.Titulo,
-                Video = unaReceta.Video,
-                pasos = await CreatePasoResponse(unaReceta.Pasos),
-                ingredientes = await CreateIngredienteRecetaResponse(unaReceta.IngredentesReceta)
-            };
-            return receta;
-        }
-        private Task<List<PasoResponse>> CreatePasoResponse (ICollection<Paso> listaPasos)
-        {
-            List<PasoResponse> pasosResponse = new();
-            foreach (Paso unPaso in listaPasos)
-            {
-                pasosResponse.Add(new PasoResponse
-                {
-                    Id= unPaso.PasoId,
-                    Orden=unPaso.Orden,
-                    Descripcion=unPaso.Descripcion,
-                    Foto=unPaso.Foto,
-                });
-            }
-            return Task.FromResult(pasosResponse);
-        }
-        private Task<List<IngredienteRecetaResponse>> CreateIngredienteRecetaResponse (ICollection<IngredienteReceta> listaIngReceta)
-        {
-            List<IngredienteRecetaResponse> ingRecetasResponse = new();
-            foreach (IngredienteReceta ingReceta in listaIngReceta)
-            {
-                ingRecetasResponse.Add(new IngredienteRecetaResponse
-                {
-                    id=ingReceta.IngredienteRecetaId,
-                    ingredienteId=ingReceta.IngredienteId,
-                    nombre = "sas",
-                    
-                });
-            }
-            return Task.FromResult(ingRecetasResponse);
-        }
-
-        private Task<DificultadResponse> CreateDificultadResponse(Dificultad dificultad)
-        {
-            return Task.FromResult( new DificultadResponse
-            {
-                Id = dificultad.DificultadId,
-                Nombre = dificultad.Nombre,
-            });
-        }
-
         // ------ Validadores ------
 
-        // Validador de todos los atributos de Receta
 
         private async Task<bool> VerifyAll(RecetaRequest recetaRequest)
         {
@@ -265,6 +194,8 @@ namespace Application.UseCases.SReceta
             //Crear validador UsuarioID (Este debería validarse por el microservicio de usuario, hay que ver como es lo del token)
             //Validador de vídeos o foto receta???? Esto tenemos que ver como implementarlo. Se supone que se sube una imagen
             //Validador de titulo
+
+            //Decidí que en receta service solo valide sus atributos(RecetaRequest), los otros los deriva a los demás servicios que devolverán una excepción!!
 
             if (!await VerifyDificultadId(recetaRequest.DificultadId)) { throw new ExceptionNotFound("No existe ninguna categoría para ese ID"); }
             if (!VerifyInt(recetaRequest.DificultadId)) { throw new ExceptionSintaxError("Formato érroneo para el id de dificultad, pruebe un entero"); }
@@ -289,12 +220,12 @@ namespace Application.UseCases.SReceta
         }
         private async Task<bool> VerifyRecetaId(Guid recetaId)
         {
-            return (await _query.GetRecetaById(recetaId) != null );
+            return (await _query.GetRecetaById(recetaId) != null);
         }
 
         private async Task<bool> VerifyDificultadId(int dificultadId)
         {
-           return (await _dificultadService.ValidateDificultadById(dificultadId));
+            return (await _dificultadService.ValidateDificultadById(dificultadId));
         }
         private async Task<bool> VerifyCategoriaRecetaId(int categoriaRecetaId)
         {
@@ -308,11 +239,11 @@ namespace Application.UseCases.SReceta
             return (await _query.GetTitleLength() > title.Length);
         }
 
-        private async Task<bool> VerifyFotoRecetaLenght (string fotoreceta)
+        private async Task<bool> VerifyFotoRecetaLenght(string fotoreceta)
         {
             return (await _query.GetFotoRecetaLength() > fotoreceta.Length);
         }
-        private async Task<bool> VerifyVideoLenght (string video)
+        private async Task<bool> VerifyVideoLenght(string video)
         {
             return (await _query.GetVideoLenght() > video.Length);
         }
@@ -347,7 +278,5 @@ namespace Application.UseCases.SReceta
                 throw new ExceptionSintaxError("Formato erróneo para el horario, pruebe usando hh:mm");
             }
         }
-
-
     }
 }
